src/data_processing/requirements.txt
demoji==1.1.0
jsonlines==3.1.0
pandas==1.4.4
pymongo==4.2.0
python-dotenv==0.21.0
tqdm==4.64.1
